src/requirements.txt
asn1tools
paho-mqtt
ecdsa